Correct int casting for python3

+++ src/problems/cantilever.py
@@ -6,12 +6,12 @@ class BoundaryConditions(base.BaseProblem):
     @staticmethod
     def get_fixed_nodes(nelx, nely, params):
         return (
-            [0, nely / 2, 0], [0, nely / 2, 1],
-            [nelx * 2 / 3, nely / 2, 0], [nelx * 2 / 3, nely / 2, 1])
+            [0, int(nely / 2), 0], [0, int(nely / 2), 1],
+            [int(nelx * 2 / 3), int(nely / 2), 0], [int(nelx * 2 / 3), int(nely / 2), 1])
 
     @staticmethod
     def set_forces(nelx, nely, f, params):
-        f[2 * ((nely + 1) * nelx + nely / 2) + 1, 0] = -1.0
+        f[2 * int((nely + 1) * nelx + nely / 2) + 1, 0] = -1.0
 
     @staticmethod
     def set_passive_elements(nelx, nely, lb, ub, params):
